Added random generation of enemies on tiles. For some reason Player sometimes lags at mid-corridor transfer. This also might cause a rabbit horde to appear, since they don't ever despawn.

+++ Unity/Assets/MyAssets/Scripts/WorldManagement.cs
@@ -79,7 +79,7 @@ public partial class WorldManagement : MonoBehaviour
     public void CalculatePlayerCorridorOffset(int x)
     {
         int playerWidth = 10;
-        PlayerCurrentXMapOffset = -((x + System.Math.Sign(x) * (TileWidth + playerWidth) / 2) / TileWidth) * calculateXDirectionCoeff(CameraLookDirection);
+        PlayerCurrentXMapOffset = -((x + System.Math.Sign(x) * ((int)TileWidth + playerWidth) / 2) / (int)TileWidth) * calculateXDirectionCoeff(CameraLookDirection);
     }
     public void ChangePlayerCoordinates(int x, int y)
     {
